fix mac compil

+++ Source/Common/AudioUIHelpers.cpp
@@ -48,7 +48,7 @@ void DecibelSliderUI::drawBG(Graphics& g)
 
 String DecibelSliderUI::getValueText() const
 {
-	float valueToShow = parameter->previewValue.isVoid() ? parameter->floatValue() : parameter->previewValue;
+	float valueToShow = parameter->previewValue.isVoid() ? parameter->floatValue() : (float)parameter->previewValue;
 	float decibelsToShow = parameter->previewValue.isVoid() ? decibelParam->decibels : DecibelsHelpers::valueToDecibels(parameter->previewValue);
 	return valueToShow == 0 ? "-inf" : String::formatted("%.1f", decibelsToShow);
 }
